# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MICROSERVICES BUILD USING GOLANG following Clean-Code Architecture. External packages used: Gin, Viper, Wire, Swag etc",
        "title": "Totality-Crew",
        "contact": {},
        "version": "2.0"
    },
    "host": "localhost:3000",
    "paths": {
        "/user/getbyid/{userId}": {
            "get": {
                "description": "LISTING USER BASED ON ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USER"
                ],
                "summary": "API FOR LISTING USER BASED ON ID",
                "operationId": "LIST-USER",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter the user id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/user/getbylist": {
            "post": {
                "description": "LISTING USERS DETAILS WITH ID'S",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "USER"
                ],
                "summary": "API FOR LISTING USERS BASED ON LIST OF ID'S",
                "operationId": "LIST-USER-LIST",
                "parameters": [
                    {
                        "description": "Enter the user id's",
                        "name": "user_details",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RequestUserIDList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.RequestUserIDList": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "utils.Response": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string"
                }
            }
        }
    }
}